+++ packages/core/src/renderables/index.ts
@@ -11,3 +11,4 @@ export * from "./ScrollBar"
 export * from "./composition/constructs"
 export * from "./composition/vnode"
 export * from "./composition/VRenderable"
+export * from "./Slot"
